+++ rollkit/template/boilerplate.go
@@ -1,0 +1,18 @@
+package template
+
+const (
+	ServerAddCommandsWithStartCmdOptions = "AddCommandsWithStartCmdOptions"
+
+	RollkitV0XStartHandler  = "rollserv.StartHandler"
+	RolltkitV1XStartHandler = "abciserver.StartHandler"
+	RollkitServerOptions    = `server.AddCommandsWithStartCmdOptions(
+			rootCmd,
+			app.DefaultNodeHome,
+			newApp, appExport,
+			server.StartCmdOptions{
+				AddFlags: func(cmd *cobra.Command) {
+					abciserver.AddFlags(cmd)
+				},
+				StartCommandHandler: abciserver.StartHandler(),
+			})`
+)